Organize things a little better
