+++ src/bin/infer.rs
@@ -16,6 +16,7 @@ Arguments:
 Options:
   -h, --help           Print help
   -m, --model          The model to use (e.g., 'bert')
+  -d, --data-dir       The path to the top-level data directory (defaults to 'data')
 ";
 
 #[derive(Debug)]
@@ -28,6 +29,9 @@ struct Args {
 
     /// The model to use
     model: Option<String>,
+
+    /// The path to the top-level data directory (defaults to 'data')
+    data_dir: Option<String>,
 }
 
 fn parse_args() -> Result<Args, pico_args::Error> {
@@ -37,6 +41,7 @@ fn parse_args() -> Result<Args, pico_args::Error> {
         help: pargs.contains(["-h", "--help"]),
         pipeline: pargs.free_from_str()?,
         model: pargs.opt_value_from_str(["-m", "--model"])?,
+        data_dir: pargs.opt_value_from_str(["-d", "--data-dir"])?,
     };
 
     Ok(args)
@@ -66,10 +71,6 @@ async fn main() -> Result<()> {
 
             Ok(model)
         })?;
-
-    let data_root = "data";
-    let task_root = format!("{}/snips", data_root);
-    let artifact_dir = format!("{}/model", task_root);
 
     let device = LibTorchDevice::Cuda(0);
 
@@ -110,7 +111,7 @@ async fn main() -> Result<()> {
         Autodiff<LibTorch>,
         sequence_classification::Model<Autodiff<LibTorch>>,
         snips::Dataset,
-    >(device, &model, &artifact_dir, input)?;
+    >(device, args.data_dir, &model, input)?;
 
     // Print out predictions for each sample
     for (i, (text, expected)) in samples.into_iter().enumerate() {
