+++ src/pipelines/text_classification/inference.rs
@@ -17,13 +17,16 @@ use super::{
 /// Define inference function
 pub fn infer<B: AutodiffBackend, M: Model<B> + 'static, D: Dataset<snips::Item> + 'static>(
     device: B::Device, // Device on which to perform computation (e.g., CPU or CUDA device)
+    data_dir: Option<String>, // The location of the top-level data directory
     model_name: &str,  // The name of the model (e.g., "bert-base-uncased")
-    artifact_dir: &str, // Directory containing model and config files
     samples: Vec<String>, // Text samples for inference
 ) -> anyhow::Result<(Tensor<B, 2>, M::Config)>
 where
     i64: std::convert::From<<B as burn::tensor::backend::Backend>::IntElem>,
 {
+    let data_dir = data_dir.unwrap_or("data".to_string());
+    let artifact_dir = format!("{}/pipelines/text-classification/{}", data_dir, model_name);
+
     // Load experiment configuration
     let config = M::Config::load(format!("{artifact_dir}/config.json").as_str())
         .map_err(|e| anyhow!("Unable to load config file: {}", e))?;
